+++ src/classes/FaceMatch.ts
@@ -6,8 +6,8 @@ export interface IFaceMatch {
 }
 
 export class FaceMatch implements IFaceMatch {
-  private _label: string
-  private _distance: number
+  private _label: string;
+  private _distance: number;
 
   constructor(label: string, distance: number) {
     this._label = label;