+++ src/classes/Point.ts
@@ -4,9 +4,9 @@ export interface IPoint {
 }
 
 export class Point implements IPoint {
-  private _x: number
+  private _x: number;
 
-  private _y: number
+  private _y: number;
 
   constructor(x: number, y: number) {
     this._x = x;