+++ src/faceExpressionNet/FaceExpressions.ts
@@ -1,13 +1,13 @@
 export const FACE_EXPRESSION_LABELS = ['neutral', 'happy', 'sad', 'angry', 'fearful', 'disgusted', 'surprised'];
 
 export class FaceExpressions {
-  public neutral = 0
-  public happy = 0
-  public sad = 0
-  public angry = 0
-  public fearful = 0
-  public disgusted = 0
-  public surprised = 0
+  public neutral = 0;
+  public happy = 0;
+  public sad = 0;
+  public angry = 0;
+  public fearful = 0;
+  public disgusted = 0;
+  public surprised = 0;
 
   constructor(probabilities: number[] | Float32Array) {
     if (probabilities.length !== 7) {