+++ src/ops/padToSquare.ts
@@ -8,38 +8,22 @@ import * as tf from '../../dist/tfjs.esm';
  * both sides of the minor dimension oof the image.
  * @returns The padded tensor with width === height.
  */
-export function padToSquare(
-  imgTensor: tf.Tensor4D,
-  isCenterImage = false,
-): tf.Tensor4D {
+export function padToSquare(imgTensor: tf.Tensor4D, isCenterImage = false): tf.Tensor4D {
   return tf.tidy(() => {
     const [height, width] = imgTensor.shape.slice(1);
-    if (height === width) {
-      return imgTensor;
-    }
-
+    if (height === width) return imgTensor;
     const dimDiff = Math.abs(height - width);
     const paddingAmount = Math.round(dimDiff * (isCenterImage ? 0.5 : 1));
     const paddingAxis = height > width ? 2 : 1;
-
     const createPaddingTensor = (paddingAmountLocal: number): tf.Tensor => {
       const paddingTensorShape = imgTensor.shape.slice();
       paddingTensorShape[paddingAxis] = paddingAmountLocal;
       return tf.fill(paddingTensorShape, 0, 'float32');
     };
-
     const paddingTensorAppend = createPaddingTensor(paddingAmount);
     const remainingPaddingAmount = dimDiff - (paddingTensorAppend.shape[paddingAxis] as number);
-
-    const paddingTensorPrepend = isCenterImage && remainingPaddingAmount
-      ? createPaddingTensor(remainingPaddingAmount)
-      : null;
-
-    const tensorsToStack = [
-      paddingTensorPrepend,
-      imgTensor,
-      paddingTensorAppend,
-    ]
+    const paddingTensorPrepend = isCenterImage && remainingPaddingAmount ? createPaddingTensor(remainingPaddingAmount) : null;
+    const tensorsToStack = [paddingTensorPrepend, imgTensor, paddingTensorAppend]
       .filter((t) => !!t)
       .map((t: tf.Tensor) => tf.cast(t, 'float32')) as tf.Tensor4D[];
     return tf.concat(tensorsToStack, paddingAxis);