+++ CHANGELOG.md
@@ -9,10 +9,11 @@
   
 ## Changelog
   
+### **HEAD -> master** 2021/09/29 mandic00@live.com
+
+- fix face matcher
+
 ### **1.5.4** 2021/09/29 mandic00@live.com
-
-
-### **origin/master** 2021/09/18 mandic00@live.com
 
 
 ### **1.5.3** 2021/09/16 mandic00@live.com